move web server logic to its own file
